requirements.txt
streamlit
google-generativeai
duckduckgo-search
pdfplumber